requirements.txt
numpy
Django==4.1.3
pandas
matplotlib==3.4.2
json==2.0.9
regex==2021.4.4
opencv-python==4.5.3.56
scikit-learn==0.24.2
pyresparser
mysql-connector-python